refactor(view): divide UserInterfaceView class into InputView and OutputView
- 사용자 입출력 화면을 InputView, OutputView로 나눔
- 예외처리는 Exception 클래스로 분리
